--- Sources/Extensions/TupleView_GetViews.swift
@@ -1,35 +1,0 @@
-//
-//  TupleView_GetViews.swift
-//
-//
-//  Created by Chris Larsen on 5/7/24.
-//  referenced from George's answer here: https://stackoverflow.com/questions/64238485/how-to-loop-over-viewbuilder-content-subviews-in-swiftui
-//
-
-import SwiftUI
-
-extension TupleView {
-    var getViews: [AnyView] {
-        makeArray(from: value)
-    }
-    
-    private struct GenericView {
-        let body: Any
-        
-        var anyView: AnyView? {
-            AnyView(_fromValue: body)
-        }
-    }
-    
-    private func makeArray<Tuple>(from tuple: Tuple) -> [AnyView] {
-        func convert(child: Mirror.Child) -> AnyView? {
-            withUnsafeBytes(of: child.value) { ptr -> AnyView? in
-                let binded = ptr.bindMemory(to: GenericView.self)
-                return binded.first?.anyView
-            }
-        }
-        
-        let tupleMirror = Mirror(reflecting: tuple)
-        return tupleMirror.children.compactMap(convert)
-    }
-}
